event resolver improvements and batch image deletion

+++ src/resolvers/event.ts
@@ -5,6 +5,7 @@ import {
 } from '@/generated/graphql.js';
 import { authCheck } from '@/middlewares/auth.js';
 import { prisma } from '@/utils/db.js';
+import { v2 as cloudinary } from 'cloudinary';
 
 const allEvents: QueryResolvers['allEvents'] = async () => {
   const events = await prisma.event.findMany();
@@ -56,9 +57,27 @@ const createEvent: MutationResolvers['createEvent'] = async (
 
 const updateEvent: MutationResolvers['updateEvent'] = async (
   _parent,
-  { input }
+  { input },
+  { req }
 ) => {
-  const event = await prisma.event.update({
+  const token = await authCheck(req);
+
+  const user = await prisma.user.findUnique({
+    where: {
+      email: token.email,
+    },
+  });
+  if (!user) throw new Error('User not found');
+
+  const event = await prisma.event.findUnique({
+    where: {
+      id: input.id,
+      creatorId: user.id,
+    },
+  });
+  if (!event) throw new Error('Unauthorized');
+
+  const updatedEvent = await prisma.event.update({
     where: {
       id: input.id,
     },
@@ -67,20 +86,49 @@ const updateEvent: MutationResolvers['updateEvent'] = async (
     },
   });
 
-  return event;
+  return updatedEvent;
 };
 
 const deleteEvent: MutationResolvers['deleteEvent'] = async (
   _parent,
-  { id }
+  { id },
+  { req }
 ) => {
-  const event = await prisma.event.delete({
+  const token = await authCheck(req);
+
+  const user = await prisma.user.findUnique({
+    where: {
+      email: token.email,
+    },
+  });
+  if (!user) throw new Error('User not found');
+
+  const event = await prisma.event.findUnique({
+    where: {
+      id,
+    },
+  });
+  if (!event) {
+    throw new Error('Event not found');
+  }
+
+  if (event.creatorId !== user.id) {
+    throw new Error('Unauthorized');
+  }
+
+  const deletedEvent = await prisma.event.delete({
     where: {
       id,
     },
   });
 
-  return event;
+  if (deletedEvent) {
+    cloudinary.api.delete_resources(
+      deletedEvent.images.map((image) => image.publicId)
+    );
+  }
+
+  return deletedEvent;
 };
 
 const resolvers: Resolvers = {
